Removed BusinessCardList and refactored the same functionality into BusinessCard.
