--- client/src/BusinessCardList/BusinessCardList.css
@@ -1,10 +1,0 @@
-.business-card-list{
-  max-width: 345px,
-}
-.business-card{
-  
-}
-.business-card-media{
-  width: 200px;
-  height: 150px;
-}
